layout_llm_web_rdk/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.5.2
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.1.0
deepspeed==0.16.9
einops==0.8.1
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.2
hjson==3.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
loguru==0.7.3
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
Naked==0.1.32
networkx==3.1
nibabel==5.2.1
ninja==1.11.1.4
nipype==1.8.6
nltk==3.9.1
numpy==1.24.4
ollama==0.5.1
onnxruntime==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.17.0
Pillow==9.5.0
protobuf==5.29.5
prov==2.0.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.0
Pygments==2.19.1
PyMuPDF==1.23.11
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.10.1
shapely==2.0.7
shellescape==3.8.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
traits==6.3.2
transformers==4.46.3
typer==0.16.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
Werkzeug==3.0.6
zipp==3.20.2
